Update RESTAURANT.py
The code is a restaurant dish vending machine, where it takes the input from use among the displayed list of items and outputs a bill and stores all the transactions locally for owner revenue calculation. The date and time have been added for proper recognition of the bill time, blinking of code and certain realistic coding styles have been added to make it look a real-time project.