Add files via upload
The notes dispenser is similar to an ATM, the notes are loaded and the requested amount is printed in the form of denominations w.r.t available cash in the machine. This code is an real time code and its fun to analyze the working. The user can load his own cash and test for different amounts.