Add files via upload
The code is an resemblance of a coin dispensing machine, the coins can be loaded on the top lines and then the user requested amount can be dispensed in terms of the available coins and denominations. This is an example of Coin dispenser project that was seen in old times.