Add files via upload
This code is an realistic attempt to explain the working of ATM in terms of cash dispensing and account handling. The code has been validated against all test cases and all scenarios. The Account holder names and passwords can also be preset before execution and also cash can be loaded for use.