Add files via upload
This project is an attempt to mimic the functionalities of a Money Bank management system, the project is included with certain functions like account creation, balance enquiry, withdraw, deposit and account closure. The ATM functionalities have also been added. This project is powered up with MongoDB, the pymongo libraries have been imported, the projects lacks GUI, rest of the things are very interesting.